+++ items/src/main/java/de/markusbordihn/adaptiveperformancetweaksitems/entity/ExperienceOrbManager.java
@@ -96,7 +96,6 @@ public class ExperienceOrbManager {
 
     // Ignore events which are already canceled
     if (event.isCanceled()) {
-      log.debug("Ignore canceled Experience Orb event {}!", event);
       return;
     }